Update Triangles.R
Changed from ncol to nrow when defining the 'upper' triangle. This is because if triangle has more rows than cols, then the na.rm=TRUE part of code will delete entries of triangle to the right of where it "thinks" the diagonal is. 

ie it doesn't correctly identify diagonal when triangle has more rows than columns